Print errstr() in glibc's error() (XCC)
Glibc has a bunch of error reporting helpers, such as error().  These
take an int errnum, usually a failed errno, and a few other things to
print out, similar to perror().  Internally, these functions do a
strerror and print out a text description of the errno.

We'd also like to print out errstr.  However, the caller is typically an
external package, such as 'grep'.  We could provide an alternative
family of error() function that takes the saved errstr, but it's
largely not worth the churn.

Most of the time, the errstr is the one that corresponds to the error,
so this gets us a 90% solution.  Often when there is no error, it was
zeroed, so the only time we'll get an old errstr was when there were
multiple errors.  Probably.

Rebuild glibc.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>