Add EPLAN9 to the error list (XCC)
This will help simplify the Go port, where we have some .go files that
are almost identical to Plan 9's, but for the different errno.

I went with one of the old error strings for EGREG to keep that whole
story rolling along.

Rebuild glibc if you want to use it.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>