mem: Don't make assertions about c->synth_buf
Old chans may have that set, even ones that aren't from this device,
though they shouldn't.  If we want to check for that, we can do it
elsewhere.

Qslab_trace was using c->synth_buf in older versions, and this check was
left behind.  Now that we aren't using it, let's not keep old references
around.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>